+++ mediation/InMobi/source/plugin/Assets/GoogleMobileAds/Mediation/InMobi/Platforms/Mediation/InMobiClientFactory.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2018 Google LLC
+// Copyright 2018 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -24,13 +24,13 @@ namespace GoogleMobileAds.Mediation.InMobi
         public static IInMobiClient InMobiInstance()
         {
             #if UNITY_EDITOR
-            return new GoogleMobileAds.Mediation.InMobi.Common.DummyClient();
+            return new GoogleMobileAds.Mediation.InMobi.Common.PlaceholderClient();
             #elif UNITY_ANDROID
             return GoogleMobileAds.Mediation.InMobi.Android.InMobiClient.Instance;
             #elif UNITY_IOS
             return GoogleMobileAds.Mediation.InMobi.iOS.InMobiClient.Instance;
             #else
-            return new GoogleMobileAds.Mediation.InMobi.Common.DummyClient();
+            return new GoogleMobileAds.Mediation.InMobi.Common.PlaceholderClient();
             #endif
         }
     }